+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## TBD
 
 - Migrate lint tooling to ESLint for both .js and .ts source files [#644](https://github.com/bugsnag/bugsnag-js/pull/644)
+- Rename `autoNotify` -> `autoDetectErrors`, and add `autoDetectUnhandledRejections` option [#648](https://github.com/bugsnag/bugsnag-js/pull/648)
 - Rename `autoCaptureSessions` -> `autoTrackSessions` and simplify validation logic [#647](https://github.com/bugsnag/bugsnag-js/pull/647)
 - Rename `report` to `event` [#646](https://github.com/bugsnag/bugsnag-js/pull/646)